Fixed Nginx Docker Image
Prior to this commit nginx was loading earlier than backend server,
due to which it was not able to reach to server even after
server starts. Maybe due to caching.
To fix this added nginx.entrypoint.sh script,
which checks is backend server is running and if not wait until it does
and then start nginx server