(Backend) Fixed bug and added multi device chat support
Multi Device Chat support -> Previous to this commit it was not possible
to chat (have websocket connection) from multiple devices for same
account. Now working fine.

Fixed Bug related to view /api/counselling/sessions
Was only returning with sessions in which direct user model was referenced
but ignored the ones in which user.counsellor was referenced, which was not
expected behaviour. Fixed